Clarify mobile/desktop DI
